move `StatisticTable` logic to `useTableFilters` hook
